+++ packages/block-library/src/form/view.js
@@ -20,6 +20,8 @@ document.querySelectorAll( 'form.wp-block-form' ).forEach( function ( form ) {
 		formData.formAction = form.action;
 		formData._ajax_nonce = wpBlockFormSettings.nonce;
 		formData.action = wpBlockFormSettings.action;
+		formData._wp_http_referer = window.location.href;
+		formData.formAction = form.action;
 
 		try {
 			const response = await fetch( wpBlockFormSettings.ajaxUrl, {