+++ packages/components/src/tabs/index.tsx
@@ -8,7 +8,12 @@ import * as Ariakit from '@ariakit/react';
  * WordPress dependencies
  */
 import { useInstanceId } from '@wordpress/compose';
-import { useLayoutEffect, useMemo, useRef } from '@wordpress/element';
+import {
+	useEffect,
+	useLayoutEffect,
+	useMemo,
+	useRef,
+} from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -44,8 +49,8 @@ function Tabs( {
 
 	const isControlled = selectedTabId !== undefined;
 
-	const { items, selectedId } = store.useState();
-	const { setSelectedId, move } = store;
+	const { items, selectedId, activeId } = store.useState();
+	const { setSelectedId, setActiveId } = store;
 
 	// Keep track of whether tabs have been populated. This is used to prevent
 	// certain effects from firing too early while tab data and relevant
@@ -154,26 +159,29 @@ function Tabs( {
 		setSelectedId,
 	] );
 
-	// In controlled mode, make sure browser focus follows the selected tab if
-	// the selection is changed while a tab is already being focused.
-	useLayoutEffect( () => {
-		if ( ! isControlled || ! selectOnMove ) {
+	useEffect( () => {
+		if ( ! isControlled ) {
 			return;
 		}
-		const currentItem = items.find( ( item ) => item.id === selectedId );
-		const activeElement = currentItem?.element?.ownerDocument.activeElement;
-		const tabsHasFocus = items.some( ( item ) => {
-			return activeElement && activeElement === item.element;
-		} );
+
+		const focusedElement =
+			items?.[ 0 ]?.element?.ownerDocument.activeElement;
 
 		if (
-			activeElement &&
-			tabsHasFocus &&
-			selectedId !== activeElement.id
+			! focusedElement ||
+			! items.some( ( item ) => focusedElement === item.element )
 		) {
-			move( selectedId );
+			return; // Return early if no tabs are focused.
 		}
-	}, [ isControlled, items, move, selectOnMove, selectedId ] );
+
+		// If, after ariakit re-computes the active tab, that tab doesn't match
+		// the currently focused tab, then we force an update to ariakit to avoid
+		// any mismatches, especially when navigating to previous/next tab with
+		// arrow keys.
+		if ( activeId !== focusedElement.id ) {
+			setActiveId( focusedElement.id );
+		}
+	}, [ activeId, isControlled, items, setActiveId ] );
 
 	const contextValue = useMemo(
 		() => ( {