+++ phpunit/tests/fonts/font-library/wpFontLibrary/getFontCollection.php
@@ -13,13 +13,7 @@
 class Tests_Fonts_WpFontLibrary_GetFontCollection extends WP_Font_Library_UnitTestCase {
 
 	public function test_should_get_font_collection() {
-		$my_font_collection_config = array(
-			'slug'        => 'my-font-collection',
-			'name'        => 'My Font Collection',
-			'description' => 'Demo about how to a font collection to your WordPress Font Library.',
-			'src'         => path_join( __DIR__, 'my-font-collection-data.json' ),
-		);
-		wp_register_font_collection( $my_font_collection_config );
+		wp_register_font_collection( 'my-font-collection', array( 'font_families' => array( 'mock' ) ) );
 		$font_collection = WP_Font_Library::get_font_collection( 'my-font-collection' );
 		$this->assertInstanceOf( 'WP_Font_Collection', $font_collection );
 	}